sylo-monorepo/apps/ai-chat-core/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
supabase
pydantic[email]